refactor(registry): split jobs API routes under 200-line policy
route.js (267) -> thin GET/POST handlers over new helpers:
- resolveEmbedding.js  profile-vs-resume embedding resolution (+ HyDE)
- matchPipeline.js     partitionFeedback, applyNegativeSignal, runMatchPipeline
                       (shared by GET and POST)
[id]/route.js (240) -> thin re-export; bodies in [id]/handlers.js.
- jobRecord.js         parseGptContent + mergeEnrichedFields (dedupes the
                       enrichment merge + safe gpt_content parse)
[id] now imports the shared getSupabase from matchingHelpers instead of
redefining its own. Public routes/behavior identical; build + existing
matchingHelpers test green.
